Fix libversion numbering
Version 0.1.7 built libgetdns.so.0 instead of libgetdns.so.1.  This
broke anything depending on the package keeping the same SONAME from
0.1.6 to 0.1.7.

The problem stems from a broken libversion in upstream's configure.ac.

This is worked around in 0.1.8 by another change to GETDNS_LIBVERSION,
but the change isn't correct either, since interfaces have been added
again (in partcular, the GETDNS_TRANSPORT_TLS_* constants have been
added, which affects the API at least, for tools trying to compile
against the newer version).

Note that the commentary in configure.ac about how to update
libversion is correct up to a point, but then breaks things badly.
This patch removes that as well.

Here is the canonical reference for how to update libversion:

https://www.gnu.org/software/libtool/manual/libtool.html#Updating-version-info