Initial version of CMake build.
This is just a basic build of the library. No options are support, and  the only builds tested are Xenial and MacOS (the latter using the Brew openssl package, and so requiring -DOPENSSL_ROOT_DIR=/usr/local/Cellar/openssl/1.0.2l option on the cmake command line).

Using the library is untested.